This Fixes two issue with the linux fbcon driver,
issue #34 for modulible mode changes on the RPI as the address and size of the console frame buffer changes
and hence the memory needs to be rememmaped and base addrees updated.

issue #35 if we are in a mode that being emulated the Flip code was assuming that the surface hight for the
screen was the same as the resolition which is not the case in a emulated mode change to the closest mode.

This commit also adds to the dubug information that can be enabled for the driver.